convert height and packet sequence from Int to Long and make Host::getConsensusState return CordaConsensusState instead of ConsensusState

+++ contracts/src/main/kotlin/jp/datachain/corda/ibc/types/Height.kt
@@ -3,4 +3,4 @@ package jp.datachain.corda.ibc.types
 import net.corda.core.serialization.CordaSerializable
 
 @CordaSerializable
-data class Height(val height: Int)
+data class Height(val height: Long)
